refactor: CardDetailPage -> AddCardPage 변경
