+++ src/component/AddCardPage/CardDetailHeader/CardDetailHeaderStyled.tsx
