+++ src/component/CardDetailView/CardDetailView.stories.tsx
@@ -12,8 +12,13 @@ type Story = StoryObj<typeof meta>;
 
 export const Primary: Story = {
   args: {
-    displayNumber: "1111-2222-••••-••••",
-    cardDate: "12/25",
-    cardOwnerName: "루루&가람",
+    creditcard: {
+      originNumber: "1111-2222-3333-4444",
+      displayNumber: "1111-2222-••••-••••",
+      cardDate: "12/25",
+      cardOwnerName: "루루&가람",
+      cardCVC: "123",
+      cardPassword: ["1", "2"],
+    },
   },
 };
